added abstractions
