+++ python/dalex/dalex/fairness/group_fairness/checks.py
@@ -1,44 +1,11 @@
 import numpy as np
-import pandas as pd
 from ..._explainer.helper import verbose_cat
+from ..basics.checks import check_parameters
+from ..basics.exceptions import ParameterCheckError
 
 
-def check_parameters(y, y_hat, protected, privileged, cutoff, verbose):
-    if not len(y) == len(y_hat) == len(protected):
-        raise ParameterCheckError("protected and explainer attributes y, y_hat must have the same lengths")
-
-    if list(np.unique(y)) != [0, 1]:
-        raise ParameterCheckError("explainer must predict binary target")
-
-    if not 0 <= y_hat <= 1:
-        raise ParameterCheckError("y_hat must have probabilistic output between 0 and 1")
-
-    if isinstance(protected, list):
-        try:
-            verbose_cat("protected list will be converted to nd.array")
-            protected = np.array(protected)
-        except:
-            ParameterCheckError("failed to convert list to nd.array")
-    elif isinstance(protected, pd.Series):
-
-        try:
-            verbose_cat("protected Series will be converted to nd.array")
-            protected = np.array(protected)
-        except:
-            ParameterCheckError("failed to convert list to nd.array")
-
-    if protected.dtype.type is not np.string_:
-        verbose_cat("protected array is not string type, converting to string ", verbose)
-        try:
-            protected = protected.astype(str)
-        except:
-            ParameterCheckError('Could not convert protected to String type')
-
-    if not isinstance(privileged, str):
-        raise ParameterCheckError('privileged parameter must be a String')
-
-    if privileged not in protected:
-        raise ParameterCheckError("privileged parameter must be in protected vector")
+def check_classifier_parameters(y, y_hat, protected, privileged, cutoff, verbose):
+    y, y_hat, protected, privileged = check_parameters(y, y_hat, protected, privileged, verbose)
 
     if isinstance(cutoff, float):
         if not 0 < cutoff < 1:
@@ -63,18 +30,3 @@ def check_epsilon(epsilon):
 
     return epsilon
 
-
-class ParameterCheckError(Exception):
-
-    def __init__(self, *args):
-        if args:
-            self.message = args[0]
-        else:
-            self.message = None
-
-    def __str__(self):
-
-        if self.message:
-            return f'Parameter Check Error, {self.message}'
-        else:
-            return 'Parameter Check Error has been raised'
